+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout-selectors.js
@@ -4,6 +4,8 @@ export const ORDER_TIME_DROPDOWN = '[data-test-id="formfield-order-time-dropdown
 export const ORDER_TIME_DROPDOWN_OPTIONS = '[data-test-id="formfield-order-time-dropdown-select"] option';
 export const USER_NOTE_INPUT = '[data-test-id="user-note"] textarea';
 export const GO_TO_PAYMENT_BUTTON = '[data-test-id="confirm-payment-submit-button"]';
+export const KNOB_BUTTON = '[id$="tabbutton-knobs"]';
+export const KNOB_CHECKOUT_DROPDOWN = '[name="Checkout Url"]';
 
 export const FIELDS = {
     mobileNumber: {
@@ -36,7 +38,7 @@ export const FIELDS = {
 export default {
     ALLERGEN_LINK,
     CHECKOUT_COMPONENT,
-    ORDER_TIME_DROPDOWN, 
+    ORDER_TIME_DROPDOWN,
     ORDER_TIME_DROPDOWN_OPTIONS,
     USER_NOTE_INPUT,
     GO_TO_PAYMENT_BUTTON,