+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-delivery.component.spec.js
@@ -2,33 +2,34 @@ import forEach from 'mocha-each';
 import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
 
 describe('f-checkout "delivery" component tests', () => {
-    beforeEach(() => {
+    before(() => {
         browser.url('?path=/story/components-organisms--checkout-component');
         browser.switchToFrame(0);
         CheckoutComponent.waitForCheckoutComponent();
     });
 
     forEach(['mobileNumber', 'addressLine1', 'addressCity', 'addressPostcode'])
-    .it.skip('each fields error message should be displayed', field => {
+    .it('each fields error message should be displayed', field => {
         // Act
-        CheckoutComponent.submit();
+        CheckoutComponent.clearCheckoutForm();
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldErrorDisplayed(field)).toBe(true);
     });
 
     forEach(['addressLine1', 'addressLine2', 'addressCity', 'addressPostcode'])
-    .it.skip('address fields should exist', field => {
+    .it('address fields should exist', field => {
         // Assert
-        expect(CheckoutComponent.doesInputFieldExist(field)).toBe(true);
+        expect(CheckoutComponent.doesFieldExist(field)).toBe(true);
     });
 
-    it.skip('should display the mandatory fields', () => {
+    it('should display the mandatory fields', () => {
         // Assert
         expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
     });
 
-    it.skip('should display a "mobileNumber" error message when a number format is incorrectly entered', () => {
+    it('should display a "mobileNumber" error message when a number format is incorrectly entered', () => {
         // Arrange
         const addressDetails = {
             mobileNumber: '+4512345678911'
@@ -36,13 +37,13 @@ describe('f-checkout "delivery" component tests', () => {
 
         // Act
         CheckoutComponent.populateCheckoutForm(addressDetails);
-        CheckoutComponent.submit();
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(true);
     });
 
-    it.skip('should not display a "mobileNumber" error message when a number is formatted correctly', () => {
+    it('should not display a "mobileNumber" error message when a number is formatted correctly', () => {
         // Arrange
         const addressDetails = {
             mobileNumber: '+4412345678911'
@@ -50,7 +51,7 @@ describe('f-checkout "delivery" component tests', () => {
 
         // Act
         CheckoutComponent.populateCheckoutForm(addressDetails);
-        CheckoutComponent.submit();
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(false);
@@ -64,11 +65,11 @@ describe('f-checkout "delivery" component tests', () => {
 
         // Act
         CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.submit();
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldTypeErrorDisplayed('addressPostcode')).toBe(true);
-    }); 
+    });
 
     it('should enable a user to submit a postcode with correct characters', () => {
         // Arrange
@@ -78,7 +79,7 @@ describe('f-checkout "delivery" component tests', () => {
 
         // Act
         CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.submit();
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldTypeErrorDisplayed('addressPostcode')).toBe(false);