+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -4,8 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-Latest (add to next release)
-------------------------------
+v0.31.0
+-------------------------------
+*January 5, 2021*
+
+
+### Changed	### Changed
+- Updated skipped tests to work with Storybook component tests.
+
+### Added
+- New `clearCheckoutForm` function added to clear the fields in tests in the `f-checkout-component.js` file.
+
 *December 30, 2020*
 
 ### Changed